# company/Ecosistemi/11-APEX-7-CORE/orchestration-layer/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OCP Local Secure Pilot",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/challenges/{operator_id}": {
      "post": {
        "summary": "Challenge",
        "operationId": "challenge_v1_auth_challenges__operator_id__post",
        "parameters": [
          {
            "name": "operator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Challenge V1 Auth Challenges  Operator Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/verify": {
      "post": {
        "summary": "Verify",
        "operationId": "verify_v1_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify V1 Auth Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows": {
      "post": {
        "summary": "Create Workflow",
        "operationId": "create_workflow_v1_workflows_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Trace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Workflow V1 Workflows Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}": {
      "get": {
        "summary": "Get Workflow",
        "operationId": "get_workflow_v1_workflows__workflow_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Workflow V1 Workflows  Workflow Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/cancel": {
      "post": {
        "summary": "Cancel Workflow",
        "operationId": "cancel_workflow_v1_workflows__workflow_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-Trace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Workflow V1 Workflows  Workflow Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/approve": {
      "post": {
        "summary": "Approve Disabled",
        "operationId": "approve_disabled_v1_workflows__workflow_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Disabled V1 Workflows  Workflow Id  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/events": {
      "get": {
        "summary": "Workflow Events",
        "operationId": "workflow_events_v1_workflows__workflow_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Workflow Events V1 Workflows  Workflow Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Live Health Live Get"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ready Health Ready Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateWorkflowRequest": {
        "properties": {
          "workflow_type": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_-]{2,63}$",
            "title": "Workflow Type"
          },
          "goal": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Goal"
          },
          "risk_hint": {
            "type": "string",
            "pattern": "^R[01]$",
            "title": "Risk Hint",
            "default": "R1"
          },
          "constraints": {
            "$ref": "#/components/schemas/WorkflowConstraints"
          },
          "skill_input": {
            "$ref": "#/components/schemas/SkillInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_type",
          "goal",
          "skill_input"
        ],
        "title": "CreateWorkflowRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SkillInput": {
        "properties": {
          "repository_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Repository Files"
          },
          "artifact_path": {
            "type": "string",
            "pattern": "^adr/[A-Za-z0-9._/-]+\\.md$",
            "title": "Artifact Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "repository_files",
          "artifact_path"
        ],
        "title": "SkillInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyRequest": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id"
          },
          "signature_b64": {
            "type": "string",
            "title": "Signature B64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge_id",
          "operator_id",
          "signature_b64"
        ],
        "title": "VerifyRequest"
      },
      "WorkflowConstraints": {
        "properties": {
          "max_tasks": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Max Tasks",
            "default": 5
          },
          "deadline_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Deadline Seconds",
            "default": 300
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 200000.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "default": 30000
          },
          "max_cost_usd": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Max Cost Usd",
            "default": 2.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkflowConstraints"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}